move match query/create query into separate tests
